rename modules.email
